Extract common parts of streaming jobs to single base class
