+++ src/main/scala/com/mozilla/telemetry/streaming/ExperimentsErrorAggregator.scala
@@ -3,18 +3,19 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 package com.mozilla.telemetry.streaming
 
-import java.sql.{Date, Timestamp}
+import java.sql.Timestamp
+
 import com.mozilla.telemetry.timeseries.SchemaBuilder
+import org.apache.spark.sql.types.StructType
 
-object ExperimentsErrorAggregator {
-  val outputPrefix = "experiment_error_aggregates/v1"
-  val queryName = "experiment_error_aggregates"
+object ExperimentsErrorAggregator extends ErrorAggregatorBase {
+  override val outputPrefix = "experiment_error_aggregates/v1"
+  override val queryName = "experiment_error_aggregates"
 
-  val defaultCountHistogramErrorsSchema = (new SchemaBuilder()).build
-  val defaultThresholdHistograms: Map[String, (List[String], List[Int])] = Map.empty
+  override val countHistogramErrorsSchema: StructType = new SchemaBuilder().build
 
-  val defaultDimensionsSchema = new SchemaBuilder()
-    .add[Timestamp]("timestamp")  // Windowed
+  override val dimensionsSchema: StructType = new SchemaBuilder()
+    .add[Timestamp]("timestamp") // Windowed
     .add[String]("submission_date_s3")
     .add[String]("channel")
     .add[String]("version")
@@ -24,7 +25,7 @@ object ExperimentsErrorAggregator {
     .add[String]("experiment_branch")
     .build
 
-  val defaultMetricsSchema = new SchemaBuilder()
+  override val metricsSchema: StructType = new SchemaBuilder()
     .add[Float]("usage_hours")
     .add[Int]("count")
     .add[Int]("main_crashes")
@@ -35,14 +36,4 @@ object ExperimentsErrorAggregator {
     .add[Int]("gmplugin_crashes")
     .add[Int]("content_shutdown_crashes")
     .build
-
-  def main(args: Array[String]): Unit = {
-    ErrorAggregator.setPrefix(outputPrefix)
-    ErrorAggregator.setQueryName(queryName)
-
-    ErrorAggregator.run(args,
-      defaultDimensionsSchema,
-      defaultMetricsSchema,
-      defaultCountHistogramErrorsSchema)
-  }
 }
