+++ src/test/scala/com/mozilla/telemetry/streaming/TestExperimentsErrorAggregator.scala
@@ -6,7 +6,6 @@ package com.mozilla.telemetry.streaming
 import java.sql.Timestamp
 
 import com.holdenkarau.spark.testing.DataFrameSuiteBase
-import org.apache.spark.sql.SparkSession
 import org.json4s.DefaultFormats
 import org.scalatest.{FlatSpec, Matchers}
 
@@ -32,9 +31,7 @@ class TestExperimentsErrorAggregator extends FlatSpec with Matchers with DataFra
         ++ contentCrashes
         ++ TestUtils.generateMainMessages(k)).map(_.toByteArray).seq
 
-    val df = ErrorAggregator.aggregate(spark.sqlContext.createDataset(messages).toDF, raiseOnError = true,
-      ExperimentsErrorAggregator.defaultDimensionsSchema, ExperimentsErrorAggregator.defaultMetricsSchema,
-      ExperimentsErrorAggregator.defaultCountHistogramErrorsSchema)
+    val df = ErrorAggregator.aggregate(spark.sqlContext.createDataset(messages).toDF, raiseOnError = true)
 
     // 1 for each experiment (there are 2), and one for a null experiment
     df.count() should be (3)
